Rename as ImportFileTest
