+++ tests/Http/CallPowerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\Http;
 
 use Exception;
 use Tests\TestCase;
